fix: minor refactor on files for components
